requirements.txt
# Requirements
netifaces==0.10.9
numpy
paramiko
protobuf==3.11.0
pyyaml
gast==0.2.2

# Dev Requirements
pytest
pytest-cov
pydocstyle
prospector
